load textures from files

+++ src/rendering/Screen.h
@@ -4,7 +4,7 @@
 
 class Screen : public Surface {
 public:
-    Screen(int width, int height, uint32_t* pixel_data) : Surface(width, height), pixel_data_(pixel_data) {}
+    Screen(uint16_t width, uint16_t height, uint32_t* pixel_data) : Surface(width, height), pixel_data_(pixel_data) {}
 
 protected:
     [[nodiscard]] uint32_t* PixelData() override {
